Extract narrative editor helpers
